backend/requirements.txt
fastapi
uvicorn
motor
groq
python-dotenv
pydantic
